+++ _Typings/gameApi/types/player.d.ts
@@ -101,7 +101,7 @@ declare module "game:type" {
          * @param paths - The paths of the materials to set as the player's renderer materials.
          * @returns
          */
-        readonly setRendererMaterials: <const T extends string[]>(paths: AssertPaths<T, "Materials">) => void
+        readonly setRendererMaterials: <const T extends readonly string[]>(paths: AssertPaths<T, "Materials">) => void
     }
 }
 
@@ -182,6 +182,6 @@ declare module "__future__" {
          * @param paths - The paths of the materials to set as the player's renderer materials.
          * @returns
          */
-        readonly setRendererMaterials: <const T extends string[]>(paths: AssertPaths<T, "Materials">) => void
+        readonly setRendererMaterials: <const T extends readonly string[]>(paths: AssertPaths<T, "Materials">) => void
     }
 }